+++ src/resources/utils/ui-constants.ts
@@ -8,7 +8,7 @@ export module UIConstants {
   export var AppKey = 'AUF';
   export var Title = 'Aurelia UI Framework';
   export var SubTitle = '';
-  export var Version = '3.0.0';
+  export var Version = '4.0.0';
 
   export var Http = {
     BaseUrl: './',